refactor: restructure web routes into modular components
- Move web route handlers from monolithic router.py to dedicated modules
- Create separate files for each route category (equipment, categories, clients, etc.)
- Implement consistent module structure across all route components
- Update route initialization in __init__.py to include all new modules
- Remove duplicate API routes for better maintainability
- Simplify router.py by making it a pure configuration file
- Ensure backward compatibility with existing endpoints
- Improve code organization for easier maintenance and future extensions
