+++ backend/web/routes/__init__.py
@@ -7,8 +7,13 @@ Each module defines routes for a specific feature or section of the application.
 from typing import List
 
 # Import routes modules
-from backend.web.routes import bookings
+from backend.web.routes import bookings, categories, clients, equipment, home, scanner
 
 __all__: List[str] = [
     'bookings',
+    'categories',
+    'clients',
+    'equipment',
+    'home',
+    'scanner',
 ]
